refactor(nicknamer): remove `reveal` module and unused code
Eliminated the `reveal` module after integrating its functionality directly into `NicknamerImpl`. Simplified the architecture by consolidating the logic, reducing redundancy, and aligning with recent refactoring efforts. Updated associated imports and traits for consistency.
